+++ include/Small/GameObjects.hpp
@@ -1,0 +1,25 @@
+/**
+ * @file GameObjects.hpp
+ * @author Alexandre Plateau (lexplt.dev@gmail.com)
+ * @brief Include all the files from the GameObjects folder
+ * @version 0.2
+ * @date 2020-04-15
+ * 
+ * @copyright Copyright (c) 2020
+ * 
+ */
+
+#ifndef sgl_small_gameobjects
+#define sgl_small_gameobjects
+
+/**
+ * @brief Contain all the code related to the game objects
+ * @details This include an Actor class and a ressource management utility
+ * 
+ */
+namespace sgl::GameObjects {}
+
+#include <Small/GameObjects/Actor.hpp>
+#include <Small/GameObjects/Manager.hpp>
+
+#endif