+++ include/Small/Widgets/Label.hpp
@@ -21,8 +21,6 @@ namespace sgl::Widgets
 {
     /**
      * @brief Basic label widget, doesn't listen to events by default
-     * @details The widget must be used inside a sgl::Widgets::Layout to be rendered
-     *          properly on screen.
      * 
      */
     class Label : public Base
@@ -59,7 +57,7 @@ namespace sgl::Widgets
         void onRender(sf::RenderTarget& screen, const sf::Transform& transform);
 
     protected:
-        sf::Text m_text;  //< Text of the label
+        sf::Text m_text;  ///< Text of the label
     };
 }