+++ source/Graphics/Node.cpp
@@ -8,58 +8,63 @@ namespace sgl::Graphics
     Node::~Node()
     {}
 
-    Node& move(float x, float y)
+    Node& Node::move(float x, float y)
     {
         m_transform.translate(x, y);
         return *this;
     }
 
-    Node& move(const sf::Vector2f& offset)
+    Node& Node::move(const sf::Vector2f& offset)
     {
         m_transform.translate(offset);
         return *this;
     }
 
-    Node& rotate(float angle)
+    Node& Node::rotate(float angle)
     {
         m_transform.rotate(angle);
         return *this;
     }
 
-    Node& rotate(float angle, float centerX, float centerY)
+    Node& Node::rotate(float angle, float centerX, float centerY)
     {
         m_transform.rotate(angle, centerX, centerY);
         return *this;
     }
 
-    Node& rotate(float angle, const sf::Vector2d& center)
+    Node& Node::rotate(float angle, const sf::Vector2f& center)
     {
         m_transform.rotate(angle, center);
         return *this;
     }
 
-    Node& scale(float scaleX, float scaleY)
+    Node& Node::scale(float scaleX, float scaleY)
     {
         m_transform.scale(scaleX, scaleY);
         return *this;
     }
 
-    Node& scale(const sf::Vector2f& factors)
+    Node& Node::scale(const sf::Vector2f& factors)
     {
         m_transform.scale(factors);
         return *this;
     }
 
-    Node& scale(float scaleX, float scaleY, float centerX, float centerY)
+    Node& Node::scale(float scaleX, float scaleY, float centerX, float centerY)
     {
         m_transform.scale(scaleX, scaleY, centerX, centerY);
         return *this;
     }
 
-    Node& scale(const sf::Vector2f& factors, const sf::Vector2f& center)
+    Node& Node::scale(const sf::Vector2f& factors, const sf::Vector2f& center)
     {
         m_transform.scale(factors, center);
         return *this;
+    }
+
+    sf::Transform& Node::getTransform()
+    {
+        return m_transform;
     }
 
     void Node::render(sf::RenderTarget& target, const sf::Transform& parent)